Move composable views into separate files
